organized the file structure
