+++ tests/test_most_active_community.py
@@ -1,4 +1,4 @@
-from most_active_community import most_active_community
+from functions.most_active_community import most_active_community
 
 
 def test_that_the_function_returns_a_string():
